refactor(app): extract useToast and useGlobalShortcuts hooks
App.tsx was the most-cited 'vibe coded' tell: a god component holding
app shell, state, layout, and a 100-line window keyboard handler. Move
the toast state/show/hide into useToast and the entire global keyboard
handler (plus the Ctrl+J defense) into useGlobalShortcuts. Behaviour is
unchanged: the listener is still mounted once and reads the latest
handlers through a ref. tsc strict + 99 tests stay green.
